Move KeyPairHolder interfaces to core-kernel
