+++ packages/core-kernel/src/contracts/shared/index.ts
@@ -2,5 +2,6 @@ export * from "./block-history-service";
 export * from "./download-block";
 export * from "./dynamic-fee";
 export * from "./forging-info";
+export * from "./key-pair-holder";
 export * from "./rounds";
 export * from "./transaction-history-service";
